+++ sonarts-core/tests/runner/sonartsServer.test.ts
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { start } from "../../src/runner/sonarts-server";
+import { start } from "../../src/runner/sonartsServer";
 import * as net from "net";
 import * as path from "path";
